Use secrets.token_bytes() as a nonce instead of secrets.token_hex() in both our encryption endpoints and when we encrypt our access request data.
When encrypting access request data, concatenate the nonce and encrypted data together, and b64 encode the entire thing to get it to the user.  Decrypting will do the reverse of this, b64decode, then separating the first 12 bytes as the nonce, and using that nonce and the supplied encryption key to decrypt the remaining message.

h/t @NevilleS