WIP cache original access request results and inputs into the node so we can reuse them for erasures.
Current access request results potentially filter out some array data, if an array field was an input field into a node.  But for erasures we need to know the exact index to erase - if an entrypoint field to a node is also a field to be masked, and it is also an array, potentially only some of those values should be masked. We need the original array data, and the inputs to the node, to know which indices we target.

Add build_paths_to_mask method which can take a field path and expand it. If it is not to array data, the data doesn't change. If it is array data, either the entire element in the array is targeted, or just certain elements if that array was an entrypoint.