Fix bug where erasure count confirmations may be returned instead of selected access request results when attempting to get access requests from the cache.
Because erasure requests are run after access requests, there normally isn't erasure results in the cache when we attempt to get access results.  However, if we were retrying a privacy request from a checkpoint that had already erased data, we could possibly return erasure results instead of access results, depending on if the access or erasure key was accessed last.