Move saas-related initialization to a startup hook.
Previously, the saas-related initialization was not taking effect for the general
webapp runtime - the saas registry would need to be reloaded from disk on the first
API invocation, which added unneeded overhead to that first API call.
With this change, the registry loading from disk is done on server initialization
and is not needed again.